+++ tests/Csla.DataPortalExtensionGenerator.Tests/DiagnosticTests.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using Microsoft.CodeAnalysis;
+using Ossendorf.Csla.DataPortalExtensionGenerator.Tests.Helper;
 
 namespace Ossendorf.Csla.DataPortalExtensionGenerator.Tests;
 public class DiagnosticTests {
