+++ tests/Csla.DataPortalExtensionGenerator.Tests/Helper/ModuleInitializer.cs
