+++ tests/Csla.DataPortalExtensionGenerator.Tests/Helper/RunResultWithIgnoreList.cs
