+++ tests/Csla.DataPortalExtensionGenerator.Tests/Helper/TestAnalyzerConfigOptions.cs
