+++ tests/Csla.DataPortalExtensionGenerator.Tests/Helper/TestAnalyzerConfigOptionsProvider.cs
